Upgrade to Storybook 6
