+++ .storybook/main.js
@@ -1,11 +1,10 @@
 module.exports = {
-  stories: ['../src/**/*.story.js'],
+  stories: ['../src/**/*.story.@(tsx|js)'],
   addons: [
     '@storybook/addon-a11y',
     '@storybook/addon-actions',
     '@storybook/addon-docs',
-    '@storybook/addon-knobs',
-    '@storybook/addon-options',
+    '@storybook/addon-knobs',,
     '@storybook/addon-storysource',
     '@storybook/addon-viewport',
   ],
