+++ package.json
@@ -62,14 +62,13 @@
     "@emotion/babel-preset-css-prop": "^10.0.9",
     "@storybook/addon-a11y": "^6.3.2",
     "@storybook/addon-actions": "^6.3.2",
-    "@storybook/addon-docs": "^5.0.10",
-    "@storybook/addon-knobs": "^5.0.10",
-    "@storybook/addon-options": "^5.0.10",
-    "@storybook/addon-storysource": "^5.0.10",
+    "@storybook/addon-docs": "^6.3.4",
+    "@storybook/addon-knobs": "^6.3.0",
+    "@storybook/addon-storysource": "^6.3.4",
     "@storybook/addon-viewport": "^6.3.2",
-    "@storybook/addons": "^5.0.10",
+    "@storybook/addons": "^6.3.4",
     "@storybook/react": "^6.3.2",
-    "@sumup/foundry": "^4.0.0-canary",
+    "@sumup/foundry": "^4.0.0",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/react-hooks": "^3.7.0",
