+++ src/components/Header/components/Title/Title.story.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { boolean, text } from '@storybook/addon-knobs/react';
+import { boolean, text } from '@storybook/addon-knobs';
 
 import { Title } from './Title';
